requirements.txt
flask
vanna[chromadb,postgres,mysql,ollama,openai]
dashscope
db-dtypes
python-dotenv
requests
psycopg2-binary